signal model and postfit plots
